requirements.txt
aiogram==3.1.1
aiohttp==3.9.3
python-dotenv==1.0.1
openai==1.14.3
requests==2.31.0
moviepy==1.0.3
imageio==2.34.1
imageio-ffmpeg==0.4.9
ffmpeg-python==0.2.0
pydantic==2.5.3
aiofiles==23.1.0
beautifulsoup4==4.12.3
yt-dlp==2024.4.9